requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0

# Computer Vision and Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
click>=8.0.0
wandb>=0.12.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Optional: Advanced ML
tensorflow>=2.8.0
keras>=2.8.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Optional: Visualization
plotly>=5.0.0
bokeh>=2.4.0

# Optional: Optimization
optuna>=2.10.0
hyperopt>=0.2.7 